+++ src/openassetio-python/tests/package/pluginSystem/resources/entryPoint/site-packages/packaged_plugin/__init__.py
@@ -1,0 +1,12 @@
+"""
+Provides a test PythonPluginSystemPlugin implemented within a package.
+"""
+
+# pylint gets upset, but this is fine due to
+# the way the plugin system loads plugins.
+# pylint: disable=import-error
+from .PackagePlugin import PackagePlugin
+
+
+# pylint: disable=invalid-name
+plugin = PackagePlugin