+++ src/openassetio-python/tests/package/pluginSystem/resources/pathC/modulePlugin.py
@@ -3,10 +3,10 @@ Provides a plugin that uses the same identifier as the ModulePlugin in
 the `pathA` folder.
 """
 
-from openassetio.pluginSystem import PythonPluginSystemPlugin
+from openassetio.pluginSystem import PythonPluginSystemManagerPlugin
 
 
-class ModulePlugin(PythonPluginSystemPlugin):
+class ModulePlugin(PythonPluginSystemManagerPlugin):
     """
     Provides an alternate implementation of the ModulePlugin
     to aid testing of path precedence.
@@ -16,6 +16,12 @@ class ModulePlugin(PythonPluginSystemPlugin):
     def identifier(cls):
         identifier = "org.openassetio.test.pluginSystem.resources.modulePlugin"
         return identifier
+
+    @classmethod
+    def interface(cls):
+        # This is nonsense, but allows us to check where this was
+        # loaded from in precedence checks.
+        return {"file": __file__}
 
 
 # pylint: disable=invalid-name