+++ src/openassetio-python/tests/package/pluginSystem/test_pythonpluginsystemmanagerimplementationfactory.py
@@ -18,95 +18,152 @@ These tests check the functionality of the plugin system based
 ManagerImplementationFactoryInterface implementation.
 """
 
-# pylint: disable=no-self-use
+# pylint: disable=no-self-use, unused-argument
 # pylint: disable=invalid-name,redefined-outer-name
 # pylint: disable=missing-class-docstring,missing-function-docstring
+# pylint: disable=use-implicit-booleaness-not-comparison
 
-import os
+import pytest
 
+from openassetio.log import ConsoleLogger
 from openassetio.pluginSystem import PythonPluginSystemManagerImplementationFactory
 
 
-class Test_PythonPluginSystemManagerImplementationFactory_init:
-    def test_plugin_factory_uses_the_expected_env_var(self):
+class Test_PythonPluginSystemManagerImplementationFactory:
+    def test_exposes_plugin_path_var_name_with_expected_value(self):
         assert (
             PythonPluginSystemManagerImplementationFactory.kPluginEnvVar
             == "OPENASSETIO_PLUGIN_PATH"
         )
 
-    def test_when_env_var_not_set_then_logs_warning(self, mock_logger, monkeypatch):
+    def test_exposes_disable_entry_points_var_name_with_expected_value(self):
+        assert (
+            PythonPluginSystemManagerImplementationFactory.kDisableEntryPointsEnvVar
+            == "OPENASSETIO_DISABLE_ENTRYPOINTS_PLUGINS"
+        )
+
+    def test_exposes_entry_point_group_with_expected_value(self):
+        assert (
+            PythonPluginSystemManagerImplementationFactory.kPackageEntryPointGroup
+            == "openassetio.manager_plugin"
+        )
+
+
+class Test_PythonPluginSystemManagerImplementationFactory_init:
+    def test_when_no_args_or_env_vars_then_entry_point_plugins_loaded(
+        self,
+        prepended_sys_path_with_entry_point_plugin,
+        entry_point_plugin_identifier,
+    ):
+        factory = PythonPluginSystemManagerImplementationFactory(ConsoleLogger())
+        assert factory.identifiers() == [entry_point_plugin_identifier]
+
+    def test_when_no_paths_and_entry_points_disabled_then_warning_logged(
+        self, mock_logger, monkeypatch
+    ):
         expected_msg = (
-            f"{PythonPluginSystemManagerImplementationFactory.kPluginEnvVar} is not set. "
-            "It is somewhat unlikely that you will find any plugins..."
+            "No search paths specified and entry point plugins are disabled, no plugins will "
+            f"load - check ${PythonPluginSystemManagerImplementationFactory.kPluginEnvVar} is set."
         )
         expected_severity = mock_logger.Severity.kWarning
 
-        if PythonPluginSystemManagerImplementationFactory.kPluginEnvVar in os.environ:
-            monkeypatch.delenv(PythonPluginSystemManagerImplementationFactory.kPluginEnvVar)
+        monkeypatch.delenv(
+            PythonPluginSystemManagerImplementationFactory.kPluginEnvVar, raising=False
+        )
 
-        factory = PythonPluginSystemManagerImplementationFactory(mock_logger)
+        factory = PythonPluginSystemManagerImplementationFactory(
+            mock_logger, disableEntryPointsPlugins=True
+        )
         # Plugins are scanned lazily when first requested
-        _ = factory.identifiers()
+        assert factory.identifiers() == []
         mock_logger.mock.log.assert_called_once_with(expected_severity, expected_msg)
 
+    def test_when_no_args_and_entry_points_disabled_env_then_entry_point_not_loaded(
+        self, prepended_sys_path_with_entry_point_plugin, monkeypatch
+    ):
+        monkeypatch.setenv(
+            PythonPluginSystemManagerImplementationFactory.kDisableEntryPointsEnvVar, "1"
+        )
+        monkeypatch.delenv(
+            PythonPluginSystemManagerImplementationFactory.kPluginEnvVar, raising=False
+        )
 
-class Test_PythonPluginSystemManagerImplementationFactory_identifiers:
-    def test_when_env_var_not_set_then_returns_empty_list(self, mock_logger, monkeypatch):
+        factory = PythonPluginSystemManagerImplementationFactory(ConsoleLogger())
+        assert factory.identifiers() == []
 
-        if PythonPluginSystemManagerImplementationFactory.kPluginEnvVar in os.environ:
-            monkeypatch.delenv(PythonPluginSystemManagerImplementationFactory.kPluginEnvVar)
+    def test_when_no_args_and_path_env_then_path_plugins_loaded(
+        self,
+        a_module_plugin_path,
+        module_plugin_identifier,
+        monkeypatch,
+    ):
+        monkeypatch.setenv(
+            PythonPluginSystemManagerImplementationFactory.kPluginEnvVar, a_module_plugin_path
+        )
+        factory = PythonPluginSystemManagerImplementationFactory(ConsoleLogger())
+        assert factory.identifiers() == [module_plugin_identifier]
 
-        factory = PythonPluginSystemManagerImplementationFactory(mock_logger)
-        identifiers = factory.identifiers()
-        # Check it is an empty list, not None, or any other value
-        # that would satisty == [] as a boolean comparison.
-        assert isinstance(identifiers, list)
-        assert len(identifiers) == 0
+    def test_when_path_arg_set_then_overrides_path_env(
+        self,
+        a_module_plugin_path,
+        a_package_plugin_path,
+        package_plugin_identifier,
+        mock_logger,
+        monkeypatch,
+    ):
+        monkeypatch.setenv(
+            PythonPluginSystemManagerImplementationFactory.kPluginEnvVar, a_module_plugin_path
+        )
 
-    def test_when_env_var_empty_then_returns_empty_list(self, mock_logger, monkeypatch):
+        factory = PythonPluginSystemManagerImplementationFactory(
+            paths=a_package_plugin_path, logger=mock_logger
+        )
+        assert factory.identifiers() == [
+            package_plugin_identifier,
+        ]
+
+    def test_when_entry_points_disable_arg_set_then_overrides_entry_points_disable_env(
+        self,
+        prepended_sys_path_with_entry_point_plugin,
+        entry_point_plugin_identifier,
+        mock_logger,
+        monkeypatch,
+    ):
+        monkeypatch.setenv(
+            PythonPluginSystemManagerImplementationFactory.kDisableEntryPointsEnvVar, "1"
+        )
+        factory = PythonPluginSystemManagerImplementationFactory(
+            disableEntryPointsPlugins=False, logger=mock_logger
+        )
+        assert factory.identifiers() == [
+            entry_point_plugin_identifier,
+        ]
+
+    def test_when_paths_empty_then_returns_empty_list(self, mock_logger):
 
         plugin_paths = ""
-        monkeypatch.setenv(
-            PythonPluginSystemManagerImplementationFactory.kPluginEnvVar, plugin_paths
-        )
-
-        factory = PythonPluginSystemManagerImplementationFactory(mock_logger)
-        identifiers = factory.identifiers()
-        assert isinstance(identifiers, list)
-        assert len(identifiers) == 0
-
-    def test_when_env_var_set_to_plugin_path_then_finds_plugins(
-        self, mock_logger, a_package_plugin_path, package_plugin_identifier, monkeypatch
-    ):
-
-        monkeypatch.setenv(
-            PythonPluginSystemManagerImplementationFactory.kPluginEnvVar, a_package_plugin_path
-        )
-
-        factory = PythonPluginSystemManagerImplementationFactory(mock_logger)
-        assert factory.identifiers() == [package_plugin_identifier,]
-
-    def test_when_paths_set_to_a_package_plugin_path_then_finds_local_plugins(
-        self, mock_logger, a_package_plugin_path, package_plugin_identifier, monkeypatch
-    ):
-
-        if PythonPluginSystemManagerImplementationFactory.kPluginEnvVar in os.environ:
-            monkeypatch.delenv(PythonPluginSystemManagerImplementationFactory.kPluginEnvVar)
-
         factory = PythonPluginSystemManagerImplementationFactory(
-            mock_logger, paths=a_package_plugin_path
+            paths=plugin_paths, logger=mock_logger
         )
-        assert factory.identifiers() == [package_plugin_identifier,]
+        assert factory.identifiers() == []
 
-    def test_when_env_var_overridden_to_a_package_plugin_path_then_finds_local_plugins(
-        self, mock_logger, a_package_plugin_path, package_plugin_identifier, monkeypatch
+    def test_when_duplicate_identifiers_path_selected_over_entry_point(
+        self,
+        prepended_sys_path_with_entry_point_plugin,
+        a_package_plugin_path,
+        package_plugin_identifier,
+        mock_logger,
     ):
-
-        monkeypatch.setenv(
-            PythonPluginSystemManagerImplementationFactory.kPluginEnvVar, "some invalid value"
-        )
-
         factory = PythonPluginSystemManagerImplementationFactory(
-            mock_logger, paths=a_package_plugin_path
+            mock_logger, paths=a_package_plugin_path, disableEntryPointsPlugins=False
         )
-        assert factory.identifiers() == [package_plugin_identifier,]
+
+        assert factory.identifiers() == [
+            package_plugin_identifier,
+        ]
+        assert a_package_plugin_path in factory.instantiate(package_plugin_identifier)["file"]
+
+
+@pytest.fixture
+def prepended_sys_path_with_entry_point_plugin(an_entry_point_package_plugin_root, monkeypatch):
+    monkeypatch.syspath_prepend(an_entry_point_package_plugin_root)